Improve output of "add" command (fixes #4)
This spits out the output of "git status" which has instructions on what
to do to resolve the conflict.

This also adds indications of how far through what you're doing you are.
So if you're adding multiple PRs, it tells you which one it's working on
out of how many.