Rework applying commits from a pr (#14)
This reworks how pmac finds commits for a PR and applies them to the
current branch. Before, it required you to do this squirrely thing in
your git config and then it used the information from that. Now it uses
the GitHub API to pull down the patch information and uses "git am" to
apply the patches.

I updated the README.

Note, this needs more testing, but I'll wait until July to do that.