Handle "git am" results better (#24, #22)
If "git am" goes to apply the commits for a PR, but they've already been
applied, "git am" returns a 0 exit code, but says in stdout "No changes
need to be made". pmac needs to catch this so it doesn't mess up the
previous commit. This fixes that. (#24)

If "git am" hits an issue and then tells the user to go fix it and the
user hits ENTER or something, then pmac should check to see if the "git
am" issue is resolved. If it's not resolved, pmac should tell the user
to go fix it. I think this fixes that, but I'm having difficulties
testing it. (#22)

This also tweaks the output. I think it's now easier to understand who's
saying what.