Fix git am conflict handling (#22)
If there's a "git am" conflict, pmac tells the user and then waits for
the user to resolve it in another terminal window.

This fixes the handling for that situation so it doesn't cause pmac to
die.

Further, it adds some mediocre handling for when the user resolves the
"git am" conflict by doing "git am --skip". pmac should hopefully notice
and not ammed the commit summary of the top-most commit.