Switch from setup.py to pyproject.toml
This drops the top/bottom caps for dev requirements. A long time ago, I
had a hard time with people using different versions of dev requirements
and writing up issues that wasted a lot of my time. I haven't had that
recently and it adds to maintenance burden, so I'm going to drop the
requirements.